now you can also update forge, not done yet!

+++ setupForge.py
@@ -12,11 +12,33 @@ import stat
 import time
 import io
 import os
+import re
 
 from colorama import deinit, reinit, Fore, Back, Style
 
 gradlePath = r"."
 
+def getVerFromBuildGradle():
+    with io.open(os.path.join(gradlePath, "build.gradle"), mode="r", encoding="utf-8") as buildGradle:
+        content = buildGradle.read()
+        pattern = re.compile(r"minecraft\s*\{.*?version.*?=.*?\"(.*?)\"", re.DOTALL | re.UNICODE)
+        matches = pattern.findall(content)
+        if not matches is None and len(matches) > 0:
+            return matches[0]
+        else:
+            return None
+
+def updateBuildGradleVer(mod, version):
+    with io.open(os.path.join(gradlePath, "src", mod, "build.gradle"), mode="r+", encoding="utf-8") as buildGradle:
+        content = buildGradle.read()
+        pattern = re.compile(r"(.*?minecraft\s*\{.*?version.*?=.*?\").*?(\".*)", re.DOTALL | re.UNICODE | re.MULTILINE)
+        matcher = pattern.match(content)
+        output = matcher.group(1) + version + matcher.group(2)
+        buildGradle.seek(0)
+        buildGradle.truncate()
+        buildGradle.flush()
+        buildGradle.write(output)
+        
 def downloadGradle():
     print("Reading JSON data... ", end='')
     response = urllib2.urlopen("http://files.minecraftforge.net/maven/net/minecraftforge/forge/json")
@@ -33,6 +55,10 @@ def downloadGradle():
         print(" [" + Style.BRIGHT + str(id) + Style.NORMAL + "] " + packages[id])
     print
     input = raw_input("\nPlease enter a preferred build number to continue > " + Style.BRIGHT)
+    
+    if os.path.isfile(os.path.join(gradlePath, "build.gradle")):
+        fileMode = os.stat(os.path.join(gradlePath, "build.gradle"))[stat.ST_MODE]
+        os.chmod(os.path.join(gradlePath, "build.gradle"), fileMode | stat.S_IWUSR | stat.S_IWGRP | stat.S_IWOTH)
     
     print(Style.NORMAL + "Downloading Forge... ", end='')
     forgeBuild = jsonData["number"][input]["mcversion"] + "-" + jsonData["number"][input]["version"]
@@ -71,6 +97,13 @@ def runGradleTask(percent, callList):
     return True
     
 def copyDefBuildGradle():
+    print("Copy build.gradle...", end="")
+    if os.path.isfile(os.path.join(gradlePath, "src/main/build.gradle")):
+        print("\rbuild.gradle already copied! Skipping.")
+        fileMode = os.stat(os.path.join(gradlePath, "build.gradle"))[stat.ST_MODE]
+        os.chmod(os.path.join(gradlePath, "build.gradle"), fileMode & ~stat.S_IWUSR & ~stat.S_IWGRP & ~stat.S_IWOTH)
+        return
+    
     shutil.copyfile(os.path.join(gradlePath, "build.gradle"), os.path.join(gradlePath, "src/main/build.gradle"))
     fileMode = os.stat(os.path.join(gradlePath, "build.gradle"))[stat.ST_MODE]
     os.chmod(os.path.join(gradlePath, "build.gradle"), fileMode & ~stat.S_IWUSR & ~stat.S_IWGRP & ~stat.S_IWOTH)
@@ -87,6 +120,7 @@ def copyDefBuildGradle():
         buildGradle.close()
  
 def setupGradle():
+    copyDefBuildGradle()
     if not runGradleTask(00, ["clean"]):
         return
     if not runGradleTask(25, ["cleanCache"]):
@@ -95,7 +129,6 @@ def setupGradle():
         return
     if not runGradleTask(75, ["eclipse"]):
         return
-    copyDefBuildGradle()
     print("\rSetup ForgeGradle... [Done]")
 
 def call():
@@ -115,4 +148,10 @@ def call():
         print(" [" + Style.BRIGHT + "1" + Style.NORMAL + "] rebuild workspace")
         print(" [" + Style.BRIGHT + "2" + Style.NORMAL + "] update ForgeGradle", end='\n\n')
         choice = raw_input("Please enter a number from above > " + Style.BRIGHT + Fore.WHITE)
-        
+        if choice == "1":
+            setupGradle()
+        elif choice == "2":
+            downloadGradle()
+            setupGradle()
+        elif choice == "3":
+            getVerFromBuildGradle()
